fix: infinite loop in PREP when prepping lane that is the one that is supposed to be loaded
Also more/less reverted a previous fix that was believed to cause an infinite loop but it was not. Basically that change moved that message to a place where it could never actually execute. This was verified by ejecting a lane and then calling PREP. Regardless it stops trying after 20 attempts, however with that code indented we no longer get the message because it is unreachable.