feat: save the position at the start of the tool change
In order to recover printing after a failure during a tool change we need to know the toolhead position before starting the tool change. This change will save the position at the start of the tool change and then provides a macro (`RESTORE_CHANGE_TOOL_POS`) that can be invoked at the end of your `RESUME` macro to restore that position.