Add SET_NEXT_SPOOL_ID command (#490)
Adds a SET_NEXT_SPOOL_ID command that can be used by a QR scanning utility to prep the AFC for the spool_id in the next load. This allows a user to load spools entirely at the machine without a separate manual step of keeping track of loaded spool numbers and entering them in mainsail.

---------

Signed-off-by: Kurt Kiefer <kekiefer@gmail.com>